create models dir, move plugins under models
