+++ aztk/error.py
@@ -28,7 +28,7 @@ class InvalidCustomScriptError(InvalidModelError):
     pass
 
 class InvalidPluginConfigurationError(InvalidModelError):
-    # from aztk.plugins import PluginDefinition
+    # from aztk.models.plugins import PluginDefinition
 
     def __init__(self, message:str, pluginDefinition = None):
         super().__init__(message)
