+++ aztk/models/plugins/plugin_definition.py
