+++ tests/plugins/test_plugin_definition.py
@@ -1,4 +1,4 @@
-from aztk.plugins import PluginDefinition, PluginPort, PluginRunTarget
+from aztk.models.plugins import PluginDefinition, PluginPort, PluginRunTarget
 
 
 def test_create_basic_plugin():
